separate value object Price
